+++ spec/spec_helper.rb
@@ -4,5 +4,6 @@ require 'webmock/rspec'
 
 require 'search_solr_tools'
 require 'search_solr_tools/helpers/solr_format'
+require 'search_solr_tools/helpers/harvest_status'
 
 GZIP_DEFLATE_IDENTITY = 'gzip;q=1.0,deflate;q=0.6,identity;q=0.3'