+++ spec/unit/harvesters/tdar_harvester_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-describe SearchSolrTools::Harvesters::Tdar do
+describe SearchSolrTools::Harvesters::Tdar, :skip => "Obsolete harvester, would need to be updated to new status handling method" do
   before :each do
     @harvester = described_class.new 'integration'
   end